+++ packages/zenn-markdown-html/src/utils/helper.ts
@@ -1,12 +1,6 @@
 import { escapeHtml } from 'markdown-it/lib/common/utils';
 import { extractYoutubeVideoParameters } from './url-matcher';
 
-export function isGithubUrl(url: string): boolean {
-  return /^https:\/\/github\.com\/([a-zA-Z0-9](-?[a-zA-Z0-9]){0,38})\/([a-zA-Z0-9](-?[a-zA-Z0-9._]){0,99})\/blob\/[^~\s:?[*^/\\]{2,}\/[\w!\-_~.*%()'"/]+(?:#L\d+(?:-L\d+)?)?$/.test(
-    url
-  );
-}
-
 function generateYoutubeHtml(videoId: string, start?: string) {
   const escapedVideoId = escapeHtml(videoId);
